Add files via upload
Homework: IoT Data Simulation

This project simulates IoT-generated data as part of a smart tracking system. It aims to provide hands-on experience in data collection and processing, essential for real-world applications like logistics, healthcare, and environmental monitoring. The simulated dataset is generated using Python in Jupyter Notebook and saved in CSV format for analysis and sharing.